Prefix template names to disambiguate them
All the templates are in a global namespace, even if they’re in a
plugin. They are not automatically disambiguated by plugin name.

This means that if you have the core PWA plugin and the Google
Analytics plugin enabled at the same time, the head snippets will be
duplicated because `head_snippet.html` will be loaded twice from the
same plugin. It’s not deterministic which one will be loaded.

Fix that by prefixing the Google Analytics plugin templates so they
are unique.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>